Chatbot_SweetTooth/Python script/requirements.txt
nltk
numpy
tflearn
json
random
pickle
tensorflow
flask
flask_cors
mysql-connector-python